Tags are not saving for Venues form Admin Page

+++ app/Http/Controllers/AdminEventController.php
@@ -2,6 +2,7 @@
 
 namespace CityHapps\Http\Controllers;
 
+use Log;
 use Illuminate\Http\Request;
 
 use CityHapps\Http\Requests;
@@ -115,8 +116,8 @@ class AdminEventController extends Controller {
       $result = Happ::find($eventParams['id']);
 
       // Process Tags
-      $this->createTags($result, Input::get('tags'));
-      
+      $tags = Input::get('tags');
+      $this->createTags($result, $tags);
       // Process Age Levels
       $age_level_data = Input::get('ageLevels');
       $result->ageLevels()->detach();
@@ -126,10 +127,13 @@ class AdminEventController extends Controller {
       }
 
       // Process Categories
-      $category_data = Input::get('categories');
-      $result->categories()->detach();
-      foreach ($category_data as $category) {
-        $result->categories()->attach($category);
+      if(Input::has('categories'))
+      {
+        $category_data = Input::get('categories');
+        $result->categories()->detach();
+        foreach ($category_data as $category) {
+          $result->categories()->attach($category);
+        }
       }
 
       $similar = $result->similar;
@@ -185,9 +189,9 @@ class AdminEventController extends Controller {
         if (!empty($tags)) {
             //Drop previous tags for this event
             $entity->tags()->detach();
-            foreach ($tags as $tag_id => $tag) {
+            foreach ($tags as $tag) {
                 if (!isset($tag["id"])) {
-                    $new_tag = new Tag(['tag_raw' => $tag["tag_raw"]]);
+                    $new_tag = new Tag(['tag_raw' => $tag["tag_raw"], 'tag_url' => $tag["tag_raw"]]);
                     $entity->tags()->save($new_tag);
                 } else {
                     //if the tag has an id, it means it was an old saved tag and we want it back
